pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "nubix"
version = "0.1.0"
description = "A modern cloud sync manager for Ubuntu using rclone"
authors = [{ name = "Nubix Contributors" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
keywords = ["cloud", "sync", "rclone", "dropbox", "google-drive", "nextcloud"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Desktop Environment",
    "Topic :: Internet :: File Transfer Protocol (FTP)",
]
dependencies = [
    "PySide6>=6.6.0",
    "secretstorage>=3.3.0",
    "schedule>=1.2.0",
    "PyYAML>=6.0.0",
    "cryptography>=42.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-qt>=4.4",
    "pytest-mock>=3.14",
    "black>=24.0",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
nubix = "main:main"

[project.gui-scripts]
nubix-gui = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["nubix*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
